+++ LICENSES-3rd-party.md
@@ -1,0 +1,39 @@
+# License overview of included 3rd party libraries
+
+The pymdtools project is licensed under the terms of the [MIT](LICENSE.md).
+
+However, pymdtools includes several third-party Open-Source libraries, 
+which are licensed under their own respective Open-Source licenses.
+
+
+
+## Libraries directly included in pymdtools
+
+### [mistune](https://github.com/lepture/mistune) by Hsiaoming Yang
+
+Copyright (c) 2014 - 2015, Hsiaoming Yang
+
+[License](https://github.com/lepture/mistune/blob/master/LICENSE): BSD 3-Clause "New" or "Revised" License
+
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
+
+* Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
+
+* Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
+
+* Neither the name of the creator nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.
+
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+### Bootstrap v2.3.2
+
+ Copyright (c) 2012 Twitter, Inc
+ Licensed under the [Apache License v2.0](http://www.apache.org/licenses/LICENSE-2.0)
+
+### Markdown renderer (for mistune)
+
+Copyright (c) 2015 by Jaroslav Kysela
+Licence WTFPL 2
