+++ README.md
@@ -1,2 +1,31 @@
-# mdtools
-Python tools to deal with markdown file
+pymdtools 🐸
+============
+
+A small python module to manipulate markdown content.
+We use this library to manage a lot of content in markdown.
+
+Features
+--------
+
+* Read Markdown
+* Normalize markdown (it is more a beautifier)
+* Convert to html with multiple converter (python markdown, python mistune)
+* Convert to pdf with converter
+* manage links inside the markdown (change update links)
+* add some instructions (in html comment format):
+  + variable or data easy to find
+  + include file (like license or header)
+  + share content between md file (define some shared paragraph)
+
+License
+-------
+
+The pymdtools is licensed under the terms of the MIT license and is available for free.
+
+## Links
+
+* [Web site](https://aimeos.org/integrations/typo3-shop-extension/)
+* [Documentation](https://aimeos.org/docs/TYPO3)
+* [Forum](https://aimeos.org/help/typo3-extension-f16/)
+* [Issue tracker](https://github.com/aimeos/aimeos-typo3/issues)
+* [Source code](https://github.com/aimeos/aimeos-typo3)
