+++ src/python/config/pep8.conf
@@ -23,3 +23,4 @@
 ###############################################################################
 [pep8]
 ignore = E302
+exclude = mistunege.py, mdrender.py
